マージ時にプレイヤーのメンバ関数が消えてしまったので改めて追加

+++ X-MEN_Team_Project/Player.cpp
@@ -339,7 +339,7 @@ const MelLib::Vector3& Player::GetDirection() const
 	return dirVector;
 }
 
-int Player::GetHp() const
+float Player::GetHp() const
 {
 	return hp;
 }
